Refactor: DTO 하위 패키지 구조 정리
