+++ src/main/java/com/sprint/mission/discodeit/dto/channel/ChannelResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.channel;
 
 import java.time.Instant;
 import java.util.List;
