+++ src/main/java/com/sprint/mission/discodeit/dto/channel/PublicChannelCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.channel;
 
 public record PublicChannelCreateRequest(
         String name
