+++ src/main/java/com/sprint/mission/discodeit/dto/channel/UpdateChannelRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.channel;
 
 import java.util.UUID;
