+++ src/main/java/com/sprint/mission/discodeit/dto/file/CreateBinaryContentRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.file;
 
 import java.util.UUID;
