+++ src/main/java/com/sprint/mission/discodeit/dto/message/CreateMessageRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.message;
 
 import java.util.List;
 import java.util.UUID;
