+++ src/main/java/com/sprint/mission/discodeit/dto/message/UpdateMessageRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.message;
 
 import java.util.UUID;
