+++ src/main/java/com/sprint/mission/discodeit/dto/status/CreateUserStatusRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.status;
 
 import java.util.UUID;
