+++ src/main/java/com/sprint/mission/discodeit/dto/user/CreateUserRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.user;
 
 public record CreateUserRequest(
         String username,
