+++ src/main/java/com/sprint/mission/discodeit/dto/user/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.user;
 
 public record LoginRequest(
         String username,
