+++ src/main/java/com/sprint/mission/discodeit/dto/user/UpdateUserRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.user;
 
 import java.util.UUID;
