+++ src/main/java/com/sprint/mission/discodeit/dto/user/UserResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.user;
 
 import java.util.UUID;
